Back-End Code/docAux-main/backend/requirements.txt
arrow==1.3.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
greenlet==3.0.3
idna==3.6
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
referencing==0.34.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.28
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
uri-template==1.3.0
urllib3==2.2.1
webcolors==1.13
Werkzeug==3.0.1
